// src/app/schemas/applicants.ts
export class Applicants {
    admited: string;
    semester: string;
    typeOfPraxis: number;
    videoUrl: string;
    userId: number;
    university: string;
    user: {
      id: number;
      name: string;
      birthdate: string;
      email: string;
      biography: string;
      documentType: string; // Español o ingles ?
      documentNumber: number;
    };
}